refactor: make hook for demo data
