+++ apps/demo/app/[...puckPath]/client.tsx
@@ -1,52 +1,18 @@
 "use client";
 
 import { Data } from "@/core/types/Config";
-import { resolveAllData } from "@/core/lib/resolve-all-data";
 import { Puck } from "@/core/components/Puck";
 import { Render } from "@/core/components/Render";
-import { useEffect, useState } from "react";
 import { Button } from "@/core/components/Button";
 import headingAnalyzer from "@/plugin-heading-analyzer/src/HeadingAnalyzer";
-import config, { initialData } from "../../config";
-
-const isBrowser = typeof window !== "undefined";
+import config from "../../config";
+import { useDemoData } from "../../lib/use-demo-data";
 
 export function Client({ path, isEdit }: { path: string; isEdit: boolean }) {
-  // unique b64 key that updates each time we add / remove components
-  const componentKey = Buffer.from(
-    `${Object.keys(config.components).join("-")}-${JSON.stringify(initialData)}`
-  ).toString("base64");
-
-  const key = `puck-demo:${componentKey}:${path}`;
-
-  const [data] = useState<Data>(() => {
-    if (isBrowser) {
-      const dataStr = localStorage.getItem(key);
-
-      if (dataStr) {
-        return JSON.parse(dataStr);
-      }
-
-      return initialData[path] || undefined;
-    }
+  const { data, resolvedData, key } = useDemoData({
+    path,
+    isEdit,
   });
-
-  // Normally this would happen on the server, but we can't
-  // do that because we're using local storage as a database
-  const [resolvedData, setResolvedData] = useState(data);
-
-  useEffect(() => {
-    if (data && !isEdit) {
-      resolveAllData(data, config).then(setResolvedData);
-    }
-  }, [data, isEdit]);
-
-  useEffect(() => {
-    if (!isEdit) {
-      const title = data?.root.props?.title || data?.root?.title;
-      document.title = title || "";
-    }
-  }, [data, isEdit]);
 
   if (isEdit) {
     return (
